feat: prod-server export handler
