+++ packages/server/prod-server/package.json
@@ -11,6 +11,28 @@
     "dev": "modern-lib build --watch",
     "test": "jest"
   },
+  "exports": {
+    ".": {
+      "types": "./dist/types/index.d.ts",
+      "jsnext:source": "./src/index.ts",
+      "default": "./dist/cjs/index.js"
+    },
+    "./handler": {
+      "types": "./dist/types/handler.d.ts",
+      "jsnext:source": "./src/handler.ts",
+      "default": "./dist/cjs/handler.js"
+    }
+  },
+  "typesVersions": {
+    "*": {
+      ".": [
+        "./dist/types/index.d.ts"
+      ],
+      "handler": [
+        "./dist/types/handler.d.ts"
+      ]
+    }
+  },
   "dependencies": {
     "@modern-js/server-core": "workspace:*",
     "@modern-js/runtime-utils": "workspace:*",
@@ -37,6 +59,18 @@
   "publishConfig": {
     "access": "public",
     "registry": "https://registry.npmjs.org/",
-    "types": "./dist/types/index.d.ts"
+    "types": "./dist/types/index.d.ts",
+    "exports": {
+      ".": {
+        "types": "./dist/types/index.d.ts",
+        "jsnext:source": "./src/index.ts",
+        "default": "./dist/cjs/index.js"
+      },
+      "./handler": {
+        "types": "./dist/types/handler.d.ts",
+        "jsnext:source": "./src/handler.ts",
+        "default": "./dist/cjs/handler.js"
+      }
+    }
   }
 }
